Adjusted JELLYFIN_TYPE default  and description
When using this template before, I understood that emby users need to fill in something into this variable, so I left it at the default value, which caused problems for both Plex and Jellyfin.
Jellyseerr fork is about Jellyfin mainly, so the default should support this. Therefore I removed the default entry and expanded the description.

Alternative way to clarify this might be to leave the variable ('JELLYFIN_TYPE') empty and only mention in the description what to enter as Emby user. I did not test how Jellyfin